Conflicts:
	code/game.cpp

+++ code/engine/render/util.h
@@ -20,12 +20,12 @@ inline int max(int a, int b, int c, int d)
 	return max(max(a,b), max(c,d));
 }
 
-inline double min(double a, double b) 
+inline float min(float a, float b) 
 { 
 	return a < b ? a : b; 
 }
 
-inline double max(double a, double b) 
+inline float max(float a, float b) 
 { 
 	return a < b ? b : a; 
 }
@@ -37,7 +37,7 @@ inline int clamp(int x, int lower, int upper)
 	return x;
 }
 
-inline double clamp(double x, double lower, double upper)
+inline float clamp(float x, float lower, float upper)
 {
 	if(x < lower) return lower;
 	if(x > upper) return upper;
